restructure VSIP reporting to add new report
